pyproject.toml
[project]
name = "mcp-obsidian"
version = "0.3.0"
description = "MCP server to work with Obsidian via the remote REST plugin with FastAPI wrapper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "mcp>=1.1.0",
 "python-dotenv>=1.0.1",
 "requests>=2.32.3",
 "sentence-transformers>=2.2.2",
 "faiss-cpu>=1.7.4",
 "numpy>=1.24.0",
 "fastapi>=0.104.0",
 "uvicorn[standard]>=0.24.0",
 "pyjwt>=2.8.0",
 "python-multipart>=0.0.6",
 "pyyaml>=6.0.1",
]
[[project.authors]]
name = "Markus Pfundstein"
email = "markus@life-electronic.nl"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.389",
]

[project.scripts]
mcp-obsidian = "mcp_obsidian:main"
